+++ src/components/Lend/LoanSearch/LoanSearchSavedSearch.vue
@@ -1,7 +1,7 @@
 <template>
 	<div>
 		<!-- eslint-disable-next-line max-len -->
-		<div class="savedSearchContainer tw-bg-brand-100 tw-border tw-border-action tw-rounded tw-p-2 tw-flex tw-gap-1 tw-mb-2">
+		<div class="tw-bg-brand-100 tw-border tw-border-action tw-rounded tw-p-2 tw-flex tw-flex-col lg:tw-flex-row tw-gap-1 tw-mb-2">
 			<div class="tw-grow">
 				<div class="tw-flex">
 					<icon-add class="tw-mr-1" />
@@ -12,50 +12,50 @@
 				<!-- eslint-disable-next-line max-len -->
 				Found the perfect filter settings? Add them to your 'Saved Searches' to easily return to the list of borrowers that meet your lending criteria.
 			</div>
-			<div style="min-width: 250px;" class="savedSearchButtonContainer tw-self-center">
+			<div style="min-width: 250px;" class="tw-self-center tw-w-full">
 				<kv-button
 					variant="secondary"
-					class="savedSearchButton"
+					class="tw-w-full"
 					@click="openModal"
 				>
 					Add to Saved Searches
 				</kv-button>
 			</div>
+			<kv-lightbox
+				title="Save this search"
+				:visible="isLightboxVisible"
+				@lightbox-closed="closeModal"
+			>
+				<template #header>
+					<icon-add />
+					<h2 class="tw-text-h2">
+						Save this search
+					</h2>
+					<h3 class="tw-text-base">
+						You can find all your saved searches under Lend in the top menu bar.
+					</h3>
+				</template>
+				<div>
+					<label
+						for="savedSearchName"
+						class="tw-text-h4 tw-block tw-pb-1 tw-text-secondary"
+					>
+						Saved Search Name
+					</label>
+					<kv-text-input
+						class="tw-w-full tw-mb-2"
+						id="savedSearchName"
+						placeholder="i.e. Women Owned Business"
+						v-model="savedSearchName"
+					/>
+					<kv-button
+						@click.native="saveSavedSearch"
+					>
+						Create Saved Search
+					</kv-button>
+				</div>
+			</kv-lightbox>
 		</div>
-		<kv-lightbox
-			title="Save this search"
-			:visible="isLightboxVisible"
-			@lightbox-closed="closeModal"
-		>
-			<template #header>
-				<icon-add />
-				<h2 class="tw-text-h2">
-					Save this search
-				</h2>
-				<h3 class="tw-text-base">
-					You can find all your saved searches under Lend in the top menu bar.
-				</h3>
-			</template>
-			<div>
-				<label
-					for="savedSearchName"
-					class="tw-text-h4 tw-block tw-pb-1 tw-text-secondary"
-				>
-					Saved Search Name
-				</label>
-				<kv-text-input
-					class="tw-w-full tw-mb-2"
-					id="savedSearchName"
-					placeholder="i.e. Women Owned Business"
-					v-model="savedSearchName"
-				/>
-				<kv-button
-					@click.native="saveSavedSearch"
-				>
-					Create Saved Search
-				</kv-button>
-			</div>
-		</kv-lightbox>
 	</div>
 </template>
 
@@ -127,21 +127,3 @@ export default {
 	}
 };
 </script>
-
-<style lang="scss" scoped>
-@import "settings";
-
-@media screen and (max-width: 733px) {
-	.savedSearchContainer {
-		flex-direction: column;
-	}
-
-	.savedSearchButton {
-		&Container {
-			width: 100%;
-		}
-
-		width: 100%;
-	}
-}
-</style>