+++ src/pages/Lend/LoanChannelCategoryPage.vue
@@ -3,30 +3,31 @@
 		class="loan-channel-page category-page"
 		:gray-background="pageLayout === 'control'"
 	>
-		<loan-channel-category-control
-			v-if="pageLayout === 'control'"
+		<component
+			:is="pageLayoutComponent"
 			:add-bundles-exp="addBundlesExp"
 		/>
-		<loan-channel-category-climate-experiment v-if="pageLayout === 'experiment'" />
 
 		<add-to-basket-interstitial />
 	</www-page>
 </template>
 
 <script>
+import { preFetchAll } from '@/util/apolloPreFetch';
 import gql from 'graphql-tag';
 import experimentVersionFragment from '@/graphql/fragments/experimentVersion.graphql';
 import updateAddToBasketInterstitial from '@/graphql/mutation/updateAddToBasketInterstitial.graphql';
 import WwwPage from '@/components/WwwFrame/WwwPage';
 import AddToBasketInterstitial from '@/components/Lightboxes/AddToBasketInterstitial';
-import LoanChannelCategoryControl from '@/pages/Lend/LoanChannelCategoryControl';
-import LoanChannelCategoryClimateExperiment from '@/pages/Lend/LoanChannelCategoryClimateExperiment';
 import {
 	getExperimentSettingCached,
 	trackExperimentVersion
 } from '@/util/experimentUtils';
 
 import experimentAssignmentQuery from '@/graphql/query/experimentAssignment.graphql';
+
+const LoanChannelCategoryControl = () => import('@/pages/Lend/LoanChannelCategoryControl');
+const LoanChannelCategoryClimateExperiment = () => import('@/pages/Lend/LoanChannelCategoryClimateExperiment');
 
 const pageQuery = gql`
 	query LoanChannelCategoryPageExperiments {
@@ -88,8 +89,6 @@ export default {
 	},
 	components: {
 		AddToBasketInterstitial,
-		LoanChannelCategoryControl,
-		LoanChannelCategoryClimateExperiment,
 		WwwPage,
 	},
 	inject: ['apollo', 'cookieStore'],
@@ -100,11 +99,15 @@ export default {
 				title: undefined,
 				description: undefined
 			},
-			pageLayout: 'control'
+			pageLayout: 'control',
+			pageLayoutComponent: null
 		};
 	},
 	apollo: {
-		preFetch(config, client) {
+		preFetch(config, client, args) {
+			const { route } = args;
+			const { params } = route;
+
 			return client.query({
 				query: pageQuery
 			}).then(() => {
@@ -114,6 +117,19 @@ export default {
 					),
 					client.query({ query: experimentAssignmentQuery, variables: { id: 'category_loan_bundles' } }),
 				]);
+			}).then(results => {
+				const experimentSettings = results.map(result => result.data.experiment);
+				const ecoLayoutIsShown = experimentSettings
+					.find(setting => setting.id === 'lend_by_category_carousel_layout')?.version === 'b';
+				if (ecoLayoutIsShown && testCategories.includes(params.category)) {
+					// prefetch experimental layout
+					return LoanChannelCategoryClimateExperiment();
+				}
+				return LoanChannelCategoryControl();
+			}).then(resolvedImport => {
+				// Call preFetch for page layout component
+				const component = resolvedImport.default;
+				return preFetchAll([component], client, args);
 			});
 		}
 	},
@@ -161,11 +177,13 @@ export default {
 						'lend_by_category_carousel_layout',
 						'EXP-ACK-357-Aug2022',
 					);
-					this.pageLayout = version === 'shown' ? 'experiment' : 'control';
+					this.pageLayout = version === 'b' ? 'experiment' : 'control';
 				}
 			} else {
 				this.pageLayout = 'control';
 			}
+			this.pageLayoutComponent = this.pageLayout === 'experiment'
+				? LoanChannelCategoryClimateExperiment : LoanChannelCategoryControl;
 		},
 		initializeAddToBasketInterstitial() {
 			this.apollo.mutate({